Drop CPU ExternalInterfaces
Move the callback storage and buildPipeline implementation into the
iree_cpu dialect itself (IREECPUAttrs.cpp).
