--- compiler/src/iree/compiler/Codegen/Dialect/CPU/ExternalInterfaces/BUILD.bazel
@@ -1,32 +1,0 @@
-# Copyright 2026 The IREE Authors
-#
-# Licensed under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-load("//build_tools/bazel:build_defs.oss.bzl", "iree_compiler_cc_library")
-
-package(
-    default_visibility = ["//visibility:public"],
-    features = ["layering_check"],
-    licenses = ["notice"],  # Apache 2.0
-)
-
-iree_compiler_cc_library(
-    name = "ExternalModels",
-    srcs = [
-        "CPUPipelineExternalModels.cpp",
-        "Interfaces.cpp",
-    ],
-    hdrs = [
-        "CPUPipelineExternalModels.h",
-        "Interfaces.h",
-    ],
-    deps = [
-        "//compiler/src/iree/compiler/Codegen/Dialect/CPU/IR:IREECPUDialect",
-        "//compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR:IREECodegenDialect",
-        "@llvm-project//llvm:Support",
-        "@llvm-project//mlir:IR",
-        "@llvm-project//mlir:Pass",
-    ],
-)
