--- compiler/src/iree/compiler/Codegen/Dialect/CPU/ExternalInterfaces/CMakeLists.txt
@@ -1,31 +1,0 @@
-################################################################################
-# Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Codegen/Dialect/CPU/ExternalInterfaces/BUILD.bazel#
-#                                                                              #
-# Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
-# CMake-only content.                                                          #
-#                                                                              #
-# To disable autogeneration for this file entirely, delete this header.        #
-################################################################################
-
-iree_add_all_subdirs()
-
-iree_cc_library(
-  NAME
-    ExternalModels
-  HDRS
-    "CPUPipelineExternalModels.h"
-    "Interfaces.h"
-  SRCS
-    "CPUPipelineExternalModels.cpp"
-    "Interfaces.cpp"
-  DEPS
-    LLVMSupport
-    MLIRIR
-    MLIRPass
-    iree::compiler::Codegen::Dialect::CPU::IR::IREECPUDialect
-    iree::compiler::Codegen::Dialect::Codegen::IR::IREECodegenDialect
-  PUBLIC
-)
-
-### BAZEL_TO_CMAKE_PRESERVES_ALL_CONTENT_BELOW_THIS_LINE ###
