--- compiler/src/iree/compiler/Codegen/Dialect/CPU/ExternalInterfaces/Interfaces.h
@@ -1,19 +1,0 @@
-// Copyright 2026 The IREE Authors
-//
-// Licensed under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#ifndef IREE_COMPILER_CODEGEN_DIALECT_CPU_EXTERNALINTERFACES_INTERFACES_H_
-#define IREE_COMPILER_CODEGEN_DIALECT_CPU_EXTERNALINTERFACES_INTERFACES_H_
-
-#include "mlir/IR/DialectRegistry.h"
-
-namespace mlir::iree_compiler {
-
-/// Registers all external interface implementations for the IREE CPU dialect.
-void registerIREECPUExternalInterfaces(DialectRegistry &registry);
-
-} // namespace mlir::iree_compiler
-
-#endif // IREE_COMPILER_CODEGEN_DIALECT_CPU_EXTERNALINTERFACES_INTERFACES_H_
