+++ compiler/src/iree/compiler/Codegen/Dialect/CPU/IR/IREECPUAttrs.td
@@ -106,13 +106,14 @@ def IREECPU_DataTiledMMAAttr
 // CPU Pipeline Attribute
 //===----------------------------------------------------------------------===//
 
-def IREECPU_PipelineAttr : AttrDef<IREECPU_Dialect, "Pipeline"> {
+def IREECPU_PipelineAttr : AttrDef<IREECPU_Dialect, "Pipeline", [
+    DeclareAttrInterfaceMethods<IREECodegen_PipelineAttrInterface>]> {
   let mnemonic = "pipeline";
   let summary = "LLVMCPU lowering pipeline identifier.";
   let description = [{
-    Identifies an LLVMCPU lowering pipeline. PipelineAttrInterface is
-    attached via an external model (CPUPipelineExternalModels.cpp). The
-    builder callback is registered by registerCodegenLLVMCPUPasses().
+    Identifies an LLVMCPU lowering pipeline. The builder callback is
+    registered by registerCodegenLLVMCPUPasses() via
+    registerCPUPipelineBuilder().
   }];
   let parameters = (ins
     EnumParameter<IREECPU_LoweringPipeline>:$value
