+++ compiler/src/iree/compiler/Codegen/Dialect/CPU/IR/IREECPUTypes.h
@@ -40,7 +40,21 @@ struct LoweringConfigLevelInfo {
 #define GET_ATTRDEF_CLASSES
 #include "iree/compiler/Codegen/Dialect/CPU/IR/IREECPUAttrs.h.inc"
 
+namespace mlir::iree_compiler {
+struct CodegenPipelineOptions;
+} // namespace mlir::iree_compiler
+
 namespace mlir::iree_compiler::IREE::CPU {
+
+/// Callback type for CPU pipeline builders. Returns success if the pipeline
+/// was handled.
+using CPUPipelineBuilder =
+    LogicalResult (*)(Attribute pipelineAttr, OpPassManager &pm,
+                      const CodegenPipelineOptions *options);
+
+/// Registers a CPU pipeline builder callback. Called from the LLVMCPU backend
+/// at pass registration time. Thread-safe (uses std::call_once).
+void registerCPUPipelineBuilder(CPUPipelineBuilder builder);
 
 /// Returns all the tiling levels as integer values.
 SmallVector<int> getTilingLevelsAsInts();
