+++ compiler/src/iree/compiler/Codegen/Interfaces/BUILD.bazel
@@ -56,7 +56,6 @@ iree_compiler_cc_library(
         ":TensorMaskingOpInterface",
         ":UKernelOpInterface",
         ":VectorizableOpInterface",
-        "//compiler/src/iree/compiler/Codegen/Dialect/CPU/ExternalInterfaces:ExternalModels",
         "//compiler/src/iree/compiler/Codegen/Dialect/GPU/ExternalInterfaces:ExternalModels",
         "//compiler/src/iree/compiler/Codegen/Dialect/Map/ExternalInterfaces:ExternalModels",
         "//compiler/src/iree/compiler/Codegen/Dialect/PCF/ExternalInterfaces:ExternalModels",
